requirements.txt
python-telegram-bot
beautifulsoup4
schedule
requests
